+++ src/HotelPosSystem.Tests/BookingListTests.cs
@@ -110,7 +110,10 @@ namespace HotelPosSystem.Tests {
 
             removeButton.Click();
             Window modalWindow = _programWithDatabase.Application.GetAllTopLevelWindows(_automation)[0];
-            Button yesButton = modalWindow.FindFirstDescendant(descendant => descendant.ByName("Yes")).AsButton();
+
+            // AutomationId explanation: https://learn.microsoft.com/en-us/dotnet/api/system.windows.forms.dialogresult?view=windowsdesktop-8.0#fields
+            Button yesButton = Utilities.GetElement(_automation, "6", modalWindow).AsButton();
+
             yesButton.Click();
 
             Label? customerName = Utilities.GetElement(_automation, "customerName2", bookingList).AsLabel();
@@ -124,7 +127,10 @@ namespace HotelPosSystem.Tests {
 
             removeButton.Click();
             Window modalWindow = _programWithDatabase.Application.GetAllTopLevelWindows(_automation)[0];
-            Button noButton = modalWindow.FindFirstDescendant(descendant => descendant.ByName("No")).AsButton();
+
+            // AutomationId explanation: https://learn.microsoft.com/en-us/dotnet/api/system.windows.forms.dialogresult?view=windowsdesktop-8.0#fields
+            Button noButton = Utilities.GetElement(_automation, "7", modalWindow).AsButton();
+
             noButton.Click();
 
             Label? customerName = Utilities.GetElement(_automation, "customerName2", bookingList).AsLabel();